Flytt korreksjonslogikk til servicelag
I InnloggetSaksbehandler ligger logikk for henting av kontonummer
og inntektsoppslag. For å forenkle testing, og for å kunne behandle
korreksjoner og refusjoner stadig mer likt, flytter vi denne
logikken til samme nivå som tilsvarende logikk for refusjoner.

Dette forenkler også ny testkode, som kommer senere.
