+++ src/test/kotlin/no/nav/arbeidsgiver/tiltakrefusjon/autorisering/InnloggetSaksbehandlerTest.kt
@@ -50,8 +50,6 @@ class InnloggetSaksbehandlerTest(
         refusjonRepository,
         korreksjonRepository,
         refusjonService,
-        inntektskomponentService,
-        kontoregisterService,
         AdGruppeTilganger(
             beslutter = true,
             korreksjon = true,
